move trace norm
