+++ toqito/matrix_props/trace_norm.py
