+++ toqito/state_metrics/__init__.py
@@ -1,5 +1,4 @@
 """Distance metrics for quantum states."""
-from toqito.state_metrics.trace_norm import trace_norm
 from toqito.state_metrics.hilbert_schmidt import hilbert_schmidt
 from toqito.state_metrics.hilbert_schmidt_inner_product import hilbert_schmidt_inner_product
 from toqito.state_metrics.helstrom_holevo import helstrom_holevo
